Added onclick admin button, now onclick messages work
Also added function elements however .map area was causing problems creating error "Objects are not valid as a React child (found: object with keys {id, name, address}). If you meant to render a collection of children, use an array instead."

next step is to look at the commented out steps in the body of code.